@@ -1,6 +1,6 @@
 {
   "name": "rc-virtual-list",
-  "version": "3.4.11",
+  "version": "3.19.0-0",
   "description": "React Virtual List Component",
   "engines": {
     "node": ">=8.x"
@@ -22,8 +22,7 @@
   "files": [
     "lib",
     "es",
-    "dist",
-    "assets/*.css"
+    "assets"
   ],
   "license": "MIT",
   "main": "./lib/index",
@@ -34,38 +33,44 @@
     "compile": "father build",
     "prepublishOnly": "npm run compile && np --no-cleanup --yolo --no-publish",
     "lint": "eslint src/ --ext .tsx,.ts",
-    "test": "father test",
+    "test": "rc-test",
     "now-build": "npm run build"
   },
   "peerDependencies": {
-    "react": "*",
-    "react-dom": "*"
+    "react": ">=16.9.0",
+    "react-dom": ">=16.9.0"
   },
   "devDependencies": {
+    "@rc-component/father-plugin": "^1.0.2",
+    "@testing-library/jest-dom": "^5.17.0",
+    "@testing-library/react": "^12.1.5",
     "@types/classnames": "^2.2.10",
     "@types/enzyme": "^3.10.5",
     "@types/jest": "^25.1.3",
-    "@types/react": "^18.0.0",
-    "@types/react-dom": "^18.0.0",
+    "@types/react": "^18.0.8",
+    "@types/react-dom": "^18.0.3",
     "@types/warning": "^3.0.0",
+    "cheerio": "1.0.0-rc.12",
     "cross-env": "^5.2.0",
-    "dumi": "^1.1.12",
+    "dumi": "^2.2.17",
     "enzyme": "^3.1.0",
-    "enzyme-adapter-react-16": "^1.0.2",
+    "enzyme-adapter-react-16": "^1.15.6",
     "enzyme-to-json": "^3.1.4",
-    "eslint": "^7.6.0",
-    "father": "^2.29.10",
+    "eslint": "^8.56.0",
+    "eslint-plugin-unicorn": "^55.0.0",
+    "father": "^4.4.0",
     "glob": "^7.1.6",
-    "np": "^5.0.3",
+    "np": "^7.5.0",
     "rc-animate": "^2.9.1",
-    "react": "^v16.9.0-alpha.0",
-    "react-dom": "^v16.9.0-alpha.0",
-    "typescript": "^3.5.2"
+    "rc-test": "^7.0.15",
+    "react": "16.14.0",
+    "react-dom": "16.14.0",
+    "typescript": "^5.0.0"
   },
   "dependencies": {
     "@babel/runtime": "^7.20.0",
     "classnames": "^2.2.6",
     "rc-resize-observer": "^1.0.0",
-    "rc-util": "^5.15.0"
+    "rc-util": "^5.36.0"
   }
 }
