@@ -1,22 +1,7 @@
-// more config: https://d.umijs.org/config
 import { defineConfig } from 'dumi';
 
 export default defineConfig({
-  title: 'rc-virtual-list',
-  favicon:
-    'https://avatars0.githubusercontent.com/u/9441414?s=200&v=4',
-  logo:
-    'https://avatars0.githubusercontent.com/u/9441414?s=200&v=4',
-  outputPath: '.doc',
-  exportStatic: {},
-  resolve: {
-    examples: ['none'],
+  themeConfig: {
+    name: 'Tree',
   },
-  styles: [
-    `
-      .markdown table {
-        width: auto !important;
-      }
-    `,
-  ]
 });
