@@ -2,16 +2,27 @@ import * as React from 'react';
 import ResizeObserver from 'rc-resize-observer';
 import classNames from 'classnames';
 
+export type InnerProps = Pick<React.HTMLAttributes<HTMLDivElement>, 'role' | 'id'>;
+
 interface FillerProps {
   prefixCls?: string;
   /** Virtual filler height. Should be `count * itemMinHeight` */
   height: number;
   /** Set offset of visible items. Should be the top of start item position */
-  offset?: number;
+  offsetY?: number;
+  offsetX?: number;
+
+  scrollWidth?: number;
 
   children: React.ReactNode;
 
   onInnerResize?: () => void;
+
+  innerProps?: InnerProps;
+
+  rtl: boolean;
+
+  extra?: React.ReactNode;
 }
 
 /**
@@ -19,7 +30,17 @@ interface FillerProps {
  */
 const Filler = React.forwardRef(
   (
-    { height, offset, children, prefixCls, onInnerResize }: FillerProps,
+    {
+      height,
+      offsetY,
+      offsetX,
+      children,
+      prefixCls,
+      onInnerResize,
+      innerProps,
+      rtl,
+      extra,
+    }: FillerProps,
     ref: React.Ref<HTMLDivElement>,
   ) => {
     let outerStyle: React.CSSProperties = {};
@@ -29,12 +50,18 @@ const Filler = React.forwardRef(
       flexDirection: 'column',
     };
 
-    if (offset !== undefined) {
-      outerStyle = { height, position: 'relative', overflow: 'hidden' };
+    if (offsetY !== undefined) {
+      // Not set `width` since this will break `sticky: right`
+      outerStyle = {
+        height,
+        position: 'relative',
+        overflow: 'hidden',
+      };
 
       innerStyle = {
         ...innerStyle,
-        transform: `translateY(${offset}px)`,
+        transform: `translateY(${offsetY}px)`,
+        [rtl ? 'marginRight' : 'marginLeft']: -offsetX,
         position: 'absolute',
         left: 0,
         right: 0,
@@ -57,8 +84,10 @@ const Filler = React.forwardRef(
               [`${prefixCls}-holder-inner`]: prefixCls,
             })}
             ref={ref}
+            {...innerProps}
           >
             {children}
+            {extra}
           </div>
         </ResizeObserver>
       </div>
