@@ -1,11 +1,13 @@
 import * as React from 'react';
-import type { SharedConfig, RenderFunc } from '../interface';
+import type { RenderFunc, SharedConfig } from '../interface';
 import { Item } from '../Item';
 
 export default function useChildren<T>(
   list: T[],
   startIndex: number,
   endIndex: number,
+  scrollWidth: number,
+  offsetX: number,
   setNodeRef: (item: T, element: HTMLElement) => void,
   renderFunc: RenderFunc<T>,
   { getKey }: SharedConfig<T>,
@@ -13,12 +15,15 @@ export default function useChildren<T>(
   return list.slice(startIndex, endIndex + 1).map((item, index) => {
     const eleIndex = startIndex + index;
     const node = renderFunc(item, eleIndex, {
-      // style: status === 'MEASURE_START' ? { visibility: 'hidden' } : {},
+      style: {
+        width: scrollWidth,
+      },
+      offsetX,
     }) as React.ReactElement;
 
     const key = getKey(item);
     return (
-      <Item key={key} setRef={ele => setNodeRef(item, ele)}>
+      <Item key={key} setRef={(ele) => setNodeRef(item, ele)}>
         {node}
       </Item>
     );
