@@ -1,18 +1,24 @@
-import { useRef } from 'react';
 import raf from 'rc-util/lib/raf';
+import { useRef } from 'react';
 import isFF from '../utils/isFirefox';
 import useOriginScroll from './useOriginScroll';
 
 interface FireFoxDOMMouseScrollEvent {
   detail: number;
-  preventDefault: Function;
+  preventDefault: VoidFunction;
 }
 
 export default function useFrameWheel(
   inVirtual: boolean,
   isScrollAtTop: boolean,
   isScrollAtBottom: boolean,
-  onWheelDelta: (offset: number) => void,
+  isScrollAtLeft: boolean,
+  isScrollAtRight: boolean,
+  horizontalScroll: boolean,
+  /***
+   * Return `true` when you need to prevent default event
+   */
+  onWheelDelta: (offset: number, horizontal: boolean) => void,
 ): [(e: WheelEvent) => void, (e: FireFoxDOMMouseScrollEvent) => void] {
   const offsetRef = useRef(0);
   const nextFrameRef = useRef<number>(null);
@@ -22,19 +28,31 @@ export default function useFrameWheel(
   const isMouseScrollRef = useRef<boolean>(false);
 
   // Scroll status sync
-  const originScroll = useOriginScroll(isScrollAtTop, isScrollAtBottom);
+  const originScroll = useOriginScroll(
+    isScrollAtTop,
+    isScrollAtBottom,
+    isScrollAtLeft,
+    isScrollAtRight,
+  );
 
-  function onWheel(event: WheelEvent) {
-    if (!inVirtual) return;
-
+  function onWheelY(e: WheelEvent, deltaY: number) {
     raf.cancel(nextFrameRef.current);
 
-    const { deltaY } = event;
+    // Do nothing when scroll at the edge, Skip check when is in scroll
+    if (originScroll(false, deltaY)) return;
+
+    // Skip if nest List has handled this event
+    const event = e as WheelEvent & {
+      _virtualHandled?: boolean;
+    };
+    if (!event._virtualHandled) {
+      event._virtualHandled = true;
+    } else {
+      return;
+    }
+
     offsetRef.current += deltaY;
     wheelValueRef.current = deltaY;
-
-    // Do nothing when scroll at the edge, Skip check when is in scroll
-    if (originScroll(deltaY)) return;
 
     // Proxy of scroll events
     if (!isFF) {
@@ -45,9 +63,59 @@ export default function useFrameWheel(
       // Patch a multiple for Firefox to fix wheel number too small
       // ref: https://github.com/ant-design/ant-design/issues/26372#issuecomment-679460266
       const patchMultiple = isMouseScrollRef.current ? 10 : 1;
-      onWheelDelta(offsetRef.current * patchMultiple);
+      onWheelDelta(offsetRef.current * patchMultiple, false);
       offsetRef.current = 0;
     });
+  }
+
+  function onWheelX(event: WheelEvent, deltaX: number) {
+    onWheelDelta(deltaX, true);
+
+    if (!isFF) {
+      event.preventDefault();
+    }
+  }
+
+  // Check for which direction does wheel do. `sx` means `shift + wheel`
+  const wheelDirectionRef = useRef<'x' | 'y' | 'sx' | null>(null);
+  const wheelDirectionCleanRef = useRef<number>(null);
+
+  function onWheel(event: WheelEvent) {
+    if (!inVirtual) return;
+
+    // Wait for 2 frame to clean direction
+    raf.cancel(wheelDirectionCleanRef.current);
+    wheelDirectionCleanRef.current = raf(() => {
+      wheelDirectionRef.current = null;
+    }, 2);
+
+    const { deltaX, deltaY, shiftKey } = event;
+
+    let mergedDeltaX = deltaX;
+    let mergedDeltaY = deltaY;
+
+    if (
+      wheelDirectionRef.current === 'sx' ||
+      (!wheelDirectionRef.current && (shiftKey || false) && deltaY && !deltaX)
+    ) {
+      mergedDeltaX = deltaY;
+      mergedDeltaY = 0;
+
+      wheelDirectionRef.current = 'sx';
+    }
+
+    const absX = Math.abs(mergedDeltaX);
+    const absY = Math.abs(mergedDeltaY);
+
+    if (wheelDirectionRef.current === null) {
+      wheelDirectionRef.current = horizontalScroll && absX > absY ? 'x' : 'y';
+    }
+
+    if (wheelDirectionRef.current === 'y') {
+      onWheelY(event, mergedDeltaY);
+    } else {
+      onWheelX(event, mergedDeltaX);
+    }
   }
 
   // A patch for firefox
