@@ -1,6 +1,11 @@
 import { useRef } from 'react';
 
-export default (isScrollAtTop: boolean, isScrollAtBottom: boolean) => {
+export default (
+  isScrollAtTop: boolean,
+  isScrollAtBottom: boolean,
+  isScrollAtLeft: boolean,
+  isScrollAtRight: boolean,
+) => {
   // Do lock for a wheel when scrolling
   const lockRef = useRef(false);
   const lockTimeoutRef = useRef(null);
@@ -18,16 +23,23 @@ export default (isScrollAtTop: boolean, isScrollAtBottom: boolean) => {
   const scrollPingRef = useRef({
     top: isScrollAtTop,
     bottom: isScrollAtBottom,
+    left: isScrollAtLeft,
+    right: isScrollAtRight,
   });
   scrollPingRef.current.top = isScrollAtTop;
   scrollPingRef.current.bottom = isScrollAtBottom;
+  scrollPingRef.current.left = isScrollAtLeft;
+  scrollPingRef.current.right = isScrollAtRight;
 
-  return (deltaY: number, smoothOffset = false) => {
-    const originScroll =
-      // Pass origin wheel when on the top
-      (deltaY < 0 && scrollPingRef.current.top) ||
-      // Pass origin wheel when on the bottom
-      (deltaY > 0 && scrollPingRef.current.bottom);
+  return (isHorizontal: boolean, delta: number, smoothOffset = false) => {
+    const originScroll = isHorizontal
+      ? // Pass origin wheel when on the left
+        (delta < 0 && scrollPingRef.current.left) ||
+        // Pass origin wheel when on the right
+        (delta > 0 && scrollPingRef.current.right) // Pass origin wheel when on the top
+      : (delta < 0 && scrollPingRef.current.top) ||
+        // Pass origin wheel when on the bottom
+        (delta > 0 && scrollPingRef.current.bottom);
 
     if (smoothOffset && originScroll) {
       // No need lock anymore when it's smooth offset from touchMove interval
